feat: [ANDROAPP-5125] move setTeiImage extensions from commons to app
