+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/adapters/BaseTeiViewHolder.kt
@@ -8,7 +8,6 @@ import org.dhis2.bindings.hasFollowUp
 import org.dhis2.bindings.paintAllEnrollmentIcons
 import org.dhis2.bindings.setAttributeList
 import org.dhis2.bindings.setStatusText
-import org.dhis2.bindings.setTeiImage
 import org.dhis2.commons.data.EnrollmentIconData
 import org.dhis2.commons.data.SearchTeiModel
 import org.dhis2.commons.date.toDateSpan
